requirements.txt
# requirements.txt
streamlit
pdfplumber
python-docx
requests
python-dotenv

